requirements.txt
pymongo
flask
Pillow==9.0.0